+++ Readme/CHANGELOG.md
@@ -34,6 +34,7 @@ If you would like to see fresh changes done since last release you can do so on 
 
     - Microsoft Edge WebView2
     - Sysinternals sigcheck
+    - Microsoft Network Realtime Inspection Service
 
   - Updated rules for: